+++ changelog.md
@@ -57,3 +57,6 @@ routes: cart
 - added route /order - which save order in localStorage
 - manage-order - get Orders from local storage
 - added canLoad guard for admin and lazy loading
+###[6.0.0]
+- added db.json with products
+- changed productService for using http